--- src/main/java/CaffeineCoder/recipic/brand/model/Brand.java
@@ -1,20 +1,0 @@
-package CaffeineCoder.recipic.brand.model;
-
-import jakarta.persistence.*;
-import java.util.List;
-
-@Entity
-public class Brand {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "brand_id")
-    private Long brandId;
-
-    @Column(name = "brand_name")
-    private String brandName;
-
-    @OneToMany(mappedBy = "brand")
-    private List<BrandIngredient> brandIngredients;
-
-}
