--- src/main/java/CaffeineCoder/recipic/brand/repository/BrandRepository.java
@@ -1,13 +1,0 @@
-package CaffeineCoder.recipic.brand.repository;
-
-import CaffeineCoder.recipic.brand.model.Brand;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Repository
-public interface BrandRepository extends JpaRepository<Brand, Long> {
-    Optional<Brand> findByBrandName(String brandName);
-}
