--- src/main/java/CaffeineCoder/recipic/brand/repository/IngredientRepository.java
@@ -1,9 +1,0 @@
-package CaffeineCoder.recipic.brand.repository;
-
-import CaffeineCoder.recipic.brand.model.Ingredient;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IngredientRepository extends JpaRepository<Ingredient, Integer> {
-}
