+++ src/main/java/CaffeineCoder/recipic/domain/brand/domain/Brand.java
@@ -1,0 +1,20 @@
+package CaffeineCoder.recipic.domain.brand.domain;
+
+import jakarta.persistence.*;
+import java.util.List;
+
+@Entity
+public class Brand {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "brand_id")
+    private Long brandId;
+
+    @Column(name = "brand_name")
+    private String brandName;
+
+    @OneToMany(mappedBy = "brand")
+    private List<BrandIngredient> brandIngredients;
+
+}
