+++ src/main/java/CaffeineCoder/recipic/domain/brand/domain/BrandIngredient.java
@@ -1,0 +1,30 @@
+package CaffeineCoder.recipic.domain.brand.domain;
+
+import jakarta.persistence.*;
+
+@Entity
+public class BrandIngredient {
+
+    @EmbeddedId
+    private BrandIngredientId id;
+
+    @ManyToOne
+    @MapsId("brandId")
+    @JoinColumn(name = "brand_id")
+    private Brand brand;
+
+    @ManyToOne
+    @MapsId("ingredientId")
+    @JoinColumn(name = "ingredient_id")
+    private Ingredient ingredient;
+
+
+    public Ingredient getIngredient() {
+        return ingredient;
+    }
+
+    public void setIngredient(Ingredient ingredient) {
+        this.ingredient = ingredient;
+    }
+
+}
