+++ src/main/java/CaffeineCoder/recipic/domain/brand/repository/BrandRepository.java
@@ -1,0 +1,13 @@
+package CaffeineCoder.recipic.domain.brand.repository;
+
+import CaffeineCoder.recipic.domain.brand.domain.Brand;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+
+@Repository
+public interface BrandRepository extends JpaRepository<Brand, Long> {
+    Optional<Brand> findByBrandName(String brandName);
+}
