+++ src/main/java/CaffeineCoder/recipic/domain/recipe/dao/BrandRepository.java
@@ -1,0 +1,9 @@
+package CaffeineCoder.recipic.domain.recipe.dao;
+
+import CaffeineCoder.recipic.domain.recipe.domain.Brand;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface BrandRepository extends JpaRepository<Brand, Long> {
+
+}
