+++ src/main/java/CaffeineCoder/recipic/domain/recipe/dao/IngredientRepository.java
@@ -1,0 +1,9 @@
+package CaffeineCoder.recipic.domain.recipe.dao;
+
+import CaffeineCoder.recipic.domain.recipe.domain.Ingredient;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
+
+}
