+++ src/main/java/CaffeineCoder/recipic/domain/recipe/dao/RecipeIngredientRepository.java
@@ -1,0 +1,8 @@
+package CaffeineCoder.recipic.domain.recipe.dao;
+
+import CaffeineCoder.recipic.domain.recipe.domain.RecipeIngredient;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecipeIngredientRepository extends JpaRepository<RecipeIngredient, Long> {
+
+}
