+++ src/main/java/CaffeineCoder/recipic/domain/recipe/dao/RecipeRepository.java
@@ -1,0 +1,8 @@
+package CaffeineCoder.recipic.domain.recipe.dao;
+
+import CaffeineCoder.recipic.domain.recipe.domain.Recipe;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RecipeRepository extends JpaRepository<Recipe, Long> {
+
+}
