+++ src/main/java/CaffeineCoder/recipic/domain/recipe/dao/UserRepository.java
@@ -1,0 +1,8 @@
+package CaffeineCoder.recipic.domain.recipe.dao;
+
+import CaffeineCoder.recipic.domain.recipe.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
