+++ src/main/java/CaffeineCoder/recipic/domain/recipe/domain/Brand.java
@@ -1,0 +1,13 @@
+package CaffeineCoder.recipic.domain.recipe.domain;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "brand")
+public class Brand {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long brandId;
+
+    private String brandName;
+}
