+++ src/main/java/CaffeineCoder/recipic/domain/recipe/domain/User.java
@@ -1,0 +1,15 @@
+package CaffeineCoder.recipic.domain.recipe.domain;
+
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "user")
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long userId;
+
+    private String email;
+    private String nickname;
+    private Integer oAuthProvider;
+}
