--- src/main/java/CaffeineCoder/recipic/recipe/dao/BrandRepository.java
@@ -1,9 +1,0 @@
-package CaffeineCoder.recipic.recipe.dao;
-
-import CaffeineCoder.recipic.recipe.domain.Brand;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-public interface BrandRepository extends JpaRepository<Brand, Long> {
-
-}
