--- src/main/java/CaffeineCoder/recipic/recipe/dao/RecipeRepository.java
@@ -1,8 +1,0 @@
-package CaffeineCoder.recipic.recipe.dao;
-
-import CaffeineCoder.recipic.recipe.domain.Recipe;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RecipeRepository extends JpaRepository<Recipe, Long> {
-
-}
