--- src/main/java/CaffeineCoder/recipic/recipe/domain/Brand.java
@@ -1,13 +1,0 @@
-package CaffeineCoder.recipic.recipe.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "brand")
-public class Brand {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long brandId;
-
-    private String brandName;
-}
